# packages/contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BeginUpload": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "maximum": 20971520.0,
            "minimum": 1.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "checksum": {
            "$ref": "#/components/schemas/Checksum"
          },
          "content_type": {
            "enum": [
              "image/jpeg",
              "image/png"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Asset Id"
          },
          "purpose": {
            "enum": [
              "IMPORT_PHOTO",
              "IMPORT_CROP",
              "EVIDENCE"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "scope": {
            "discriminator": {
              "mapping": {
                "STORE": "#/components/schemas/StoreScope",
                "TENANT": "#/components/schemas/TenantScope"
              },
              "propertyName": "scope_kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/TenantScope"
              },
              {
                "$ref": "#/components/schemas/StoreScope"
              }
            ],
            "title": "Scope"
          }
        },
        "required": [
          "scope",
          "purpose",
          "content_type",
          "byte_size",
          "checksum"
        ],
        "title": "BeginUpload",
        "type": "object"
      },
      "Checksum": {
        "additionalProperties": false,
        "properties": {
          "algorithm": {
            "const": "SHA256",
            "title": "Algorithm",
            "type": "string"
          },
          "value": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "value"
        ],
        "title": "Checksum",
        "type": "object"
      },
      "ConnectionCreate": {
        "additionalProperties": false,
        "properties": {
          "alias": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Alias",
            "type": "string"
          },
          "kind": {
            "enum": [
              "TVT_ACCOUNT",
              "TVT_DEVICE",
              "TYCO_ACCOUNT"
            ],
            "title": "Kind",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "site": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Site",
            "type": "string"
          },
          "store_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "title": "Store Ids",
            "type": "array"
          },
          "username": {
            "format": "password",
            "maxLength": 512,
            "minLength": 1,
            "title": "Username",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "kind",
          "alias",
          "site",
          "username",
          "password"
        ],
        "title": "ConnectionCreate",
        "type": "object"
      },
      "ConnectionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConnectionView"
            },
            "title": "Items",
            "type": "array"
          },
          "selected_tenant_id": {
            "format": "uuid",
            "title": "Selected Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "items",
          "selected_tenant_id"
        ],
        "title": "ConnectionList",
        "type": "object"
      },
...
      "ConnectionView": {
        "properties": {
          "alias": {
            "title": "Alias",
            "type": "string"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "TVT_ACCOUNT",
              "TVT_DEVICE",
              "TYCO_ACCOUNT"
            ],
            "title": "Kind",
            "type": "string"
          },
          "last_success": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success"
          },
          "site": {
            "title": "Site",
            "type": "string"
          },
          "status": {
            "enum": [
              "NOT_VERIFIED",
              "DISCONNECTED"
            ],
            "title": "Status",
            "type": "string"
          },
          "store_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Store Ids",
            "type": "array"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "kind",
          "alias",
          "site",
          "status",
          "last_success",
          "generation",
          "store_ids"
        ],
        "title": "ConnectionView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoreList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StoreView"
            },
            "title": "Items",
            "type": "array"
          },
          "selected_tenant_id": {
            "format": "uuid",
            "title": "Selected Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "items",
          "selected_tenant_id"
        ],
        "title": "StoreList",
        "type": "object"
      },
      "StoreScope": {
        "additionalProperties": false,
        "properties": {
          "scope_kind": {
            "const": "STORE",
            "default": "STORE",
            "title": "Scope Kind",
            "type": "string"
          },
          "store_id": {
            "format": "uuid",
            "title": "Store Id",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "store_id"
        ],
        "title": "StoreScope",
        "type": "object"
      },
      "StoreView": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "timezone",
          "active"
        ],
        "title": "StoreView",
        "type": "object"
      },
      "TenantScope": {
        "additionalProperties": false,
        "properties": {
          "scope_kind": {
            "const": "TENANT",
            "default": "TENANT",
            "title": "Scope Kind",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id"
        ],
        "title": "TenantScope",
        "type": "object"
      },
...
      "UploadSession": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "format": "uuid",
            "title": "Asset Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "max_bytes": {
            "maximum": 20971520.0,
            "minimum": 1.0,
            "title": "Max Bytes",
            "type": "integer"
          },
          "upload_path": {
            "maxLength": 200,
            "title": "Upload Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "asset_id",
          "upload_path",
          "expires_at",
          "max_bytes"
        ],
        "title": "UploadSession",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets": {
      "post": {
        "operationId": "begin_upload_api_v1_assets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Begin Upload",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/v1/connections": {
      "get": {
        "operationId": "list_connections_api_v1_connections_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Connections",
        "tags": [
          "connections"
        ]
      },
      "post": {
        "operationId": "create_connection_api_v1_connections_post",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Connection",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/api/v1/stores": {
      "get": {
        "operationId": "stores_api_v1_stores_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stores",
        "tags": [
          "stores"
        ]
      }
    },
...
  }
}